NMS-12092 address PR comments: move tests into GraphRestServiceIT
